added graph class, assembler and minheap

+++ src/map.cpp
@@ -36,9 +36,9 @@ void Map::load(const string &filePath)
 
 void Map::findHarbors()
 {
-    for (int i = 0; i < matrix.size(); ++i)
+    for (size_t i = 0; i < matrix.size(); ++i)
     {
-        for (int j = 0; j < matrix[i].size(); ++j)
+        for (size_t j = 0; j < matrix[i].size(); ++j)
         {
             char c = matrix[i][j];
             if (c >= '1' && c <= '9')
